content_automation_project/requirements.txt
# Content Automation Project Dependencies (full desktop GUI + all pipelines).
#
# Docker image uses requirements-docker.txt only — see Dockerfile.webapp.

# Gemini API libraries
google-genai>=0.1.0
google-generativeai>=0.3.0

# PDF processing
PyMuPDF>=1.23.0  # Primary PDF library for text extraction (better quality than PyPDF2)
PyPDF2>=3.0.0  # Fallback PDF library

# GUI framework
customtkinter>=5.2.0
Pillow>=9.3.0

# Word document generation
python-docx>=1.1.0

# HTTP requests for DeepSeek API
requests>=2.31.0
python-dotenv>=1.0.0

# RAG for Reference Change: embeddings + similarity (optional; fallback to single-call if missing)
sentence-transformers>=2.2.0
numpy>=1.24.0

# Web admin app (FastAPI, jobs, Celery + Redis worker — same stack as Docker deployment)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1,<4.1
redis>=5.0.1
celery[redis]>=5.3.0
jinja2>=3.1.3
jdatetime>=4.1.0

# Standard libraries (included with Python, listed for reference)
# csv - built-in
# asyncio - built-in
# wave - built-in
# logging - built-in
# os - built-in
# typing - built-in
# datetime - built-in
# json - built-in
# threading - built-in

